+++ poptorch/source/PoptorchSymbols.cpp
@@ -33,6 +33,7 @@ c10::Symbol amin;                        // NOLINT
 c10::Symbol pow_;                        // NOLINT
 c10::Symbol sigmoid_;                    // NOLINT
 c10::Symbol tanh_;                       // NOLINT
+c10::Symbol scatter_add_;                // NOLINT
 
 // clang-format off
 __attribute__((constructor(SYMBOL_INIT_PRIORITY)))
@@ -64,6 +65,7 @@ static void initializeAtenSymbols() {
   SYMBOL_INIT(aten, pow_)
   SYMBOL_INIT(aten, sigmoid_)
   SYMBOL_INIT(aten, tanh_)
+  SYMBOL_INIT(aten, scatter_add_)
 }
 
 } // namespace c10::aten