+++ poptorch/source/PoptorchSymbols.hpp
@@ -31,6 +31,7 @@ extern c10::Symbol amin;                        // NOLINT
 extern c10::Symbol pow_;                        // NOLINT
 extern c10::Symbol sigmoid_;                    // NOLINT
 extern c10::Symbol tanh_;                       // NOLINT
+extern c10::Symbol scatter_add_;                // NOLINT
 } // namespace c10::aten
 
 namespace poptorch {